pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.poetry]
name = "simple-kafka-producer"
version = "0.1.0"
description = "A Lambda function that posts messages to a Kafka topic."
authors = ["Edward Niescior <edniescior@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "kafka_producer" },
    { include = "layers" },
]


[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^5.0"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^22.12.6"
flake8-builtins = "^2.0.1"
flake8-comprehensions = "^3.10.1"
flake8-eradicate = "^1.4.0"
flake8-implicit-str-concat = "^0.3.0"
flake8-print = "^5.0.0"
isort = "^5.11.3"
pep8-naming = "^0.13.3"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
